+++ src/theme/components/hero.ts
@@ -85,6 +85,8 @@ const getSecondaryVariantStyles = (bgColor: string = "") => {
 // Variant styling
 const primary = {
   alignItems: "center",
+  backgroundSize: "cover",
+  backgroundPosition: "center",
   display: "flex",
   flexFlow: {
     base: "column nowrap",
@@ -160,6 +162,8 @@ const tertiary = {
 };
 const campaign = {
   alignItems: "center",
+  backgroundSize: "cover",
+  backgroundPosition: "center",
   display: "flex",
   justifyContent: "center",
   marginBottom: ["0", "0", "xxl"],